chore: rename test folder
