+++ agent-control/tests/k8s/agent_control_cli/install_agent_control.rs
