+++ agent-control/tests/k8s/mod.rs
@@ -1,4 +1,4 @@
-mod cli_dynamic_objects;
+mod agent_control_cli;
 mod client;
 mod garbage_collector;
 mod scenarios;
